fix(@h6s/table): Arrange code
