+++ packages/table/src/react/useTable.stories.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 
-import { TableInstance } from '..'
 import { paymentDataset } from '../mocks/payments.mock'
 import { paymentsTableRendererModel } from '../mocks/paymentsTableRendererModel.mock'
+import { TableInstance } from '../types/table'
 import { objectEntries } from '../utils/object'
 import { useTable } from './useTable'
